master

+++ src/main/java/volta/cars/project/repository/ReservationRepository.java
@@ -12,6 +12,6 @@ public interface ReservationRepository extends CrudRepository<ReservationModel, 
     //Como hacer un Post 
     @Modifying
     @Query("INSERT INTO RESERVATIONS (USER_ID, HOST_ID, PRICE, LATITUDE, LONGITUDE, RESERVATION_DATE,RESERVATION_HOUR) VALUES (:userId, :hostId, :price, :latitude, :longitude, :reservationDate, :reservationHour)")
-    public void createReservation(int userId, int hostId, float price, float latitude, float longitude, String reservationDate, String reservationHour);
+    public void createReservation(Long userId, Long hostId, float price, float latitude, float longitude, String reservationDate, String reservationHour);
 }
